Working on the routes to create a ticket.
I have to think about how I'm going to structure my SQL database before I fix the routes.
I was thinking to write a query to join the tickets table with the events table by the event id and then check how many tickets are left to be created for that event.

I am not sure yet. Going to work on that this week.